+++ core/src/jmh/java/io/grpc/internal/SerializingExecutorBenchmark.java
@@ -17,6 +17,7 @@
 package io.grpc.internal;
 
 import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Phaser;
 import java.util.concurrent.TimeUnit;
@@ -39,7 +40,7 @@ import org.openjdk.jmh.annotations.TearDown;
 @State(Scope.Thread)
 public class SerializingExecutorBenchmark {
 
-  private ExecutorService executorService;
+  private ExecutorService executorService = Executors.newSingleThreadExecutor();
   private Executor executor = new SerializingExecutor(executorService);
 
   private static class IncrRunnable implements Runnable {
