+++ netty/build.gradle
@@ -29,3 +29,9 @@ project.sourceSets {
     }
 }
 
+
+jmh {
+    // Workaround
+    // https://github.com/melix/jmh-gradle-plugin/issues/97#issuecomment-316664026
+    includeTests = true
+}
